La Corrida/requirements.txt
botbuilder-integration-aiohttp>=4.11.0
python-telegram-bot
